refactor index read queries & change method names
no code changes
