+++ happs/observation/zomes/process/code/src/process_requests.rs
@@ -15,9 +15,9 @@ use hdk_graph_helpers::{
         update_record,
         delete_record,
     },
-    links::{
-        get_links_and_load_entry_data,
-        get_remote_links_and_load_entry_data,
+    local_indexes::{
+        query_direct_index_with_foreign_key,
+        query_direct_remote_index_with_foreign_key,
     },
     rpc::{
         RemoteEntryLinkResponse,
@@ -157,19 +157,19 @@ fn handle_query_processes(params: &QueryParams) -> ZomeApiResult<Vec<Response>> 
 
     match &params.inputs {
         Some(inputs) => {
-            entries_result = get_links_and_load_entry_data(inputs, EVENT_INPUT_OF_LINK_TYPE, EVENT_INPUT_OF_LINK_TAG);
+            entries_result = query_direct_index_with_foreign_key(inputs, EVENT_INPUT_OF_LINK_TYPE, EVENT_INPUT_OF_LINK_TAG);
         },
         _ => (),
     };
     match &params.outputs {
         Some(outputs) => {
-            entries_result = get_links_and_load_entry_data(outputs, EVENT_OUTPUT_OF_LINK_TYPE, EVENT_OUTPUT_OF_LINK_TAG);
+            entries_result = query_direct_index_with_foreign_key(outputs, EVENT_OUTPUT_OF_LINK_TYPE, EVENT_OUTPUT_OF_LINK_TAG);
         },
         _ => (),
     };
     match &params.committed_inputs {
         Some(committed_inputs) => {
-            entries_result = get_remote_links_and_load_entry_data(
+            entries_result = query_direct_remote_index_with_foreign_key(
                 committed_inputs, COMMITMENT_BASE_ENTRY_TYPE,
                 COMMITMENT_INPUT_OF_LINK_TYPE, COMMITMENT_INPUT_OF_LINK_TAG,
             );
@@ -178,7 +178,7 @@ fn handle_query_processes(params: &QueryParams) -> ZomeApiResult<Vec<Response>> 
     };
     match &params.committed_outputs {
         Some(committed_outputs) => {
-            entries_result = get_remote_links_and_load_entry_data(
+            entries_result = query_direct_remote_index_with_foreign_key(
                 committed_outputs, COMMITMENT_BASE_ENTRY_TYPE,
                 COMMITMENT_OUTPUT_OF_LINK_TYPE, COMMITMENT_OUTPUT_OF_LINK_TAG,
             );
@@ -187,7 +187,7 @@ fn handle_query_processes(params: &QueryParams) -> ZomeApiResult<Vec<Response>> 
     };
     match &params.intended_inputs {
         Some(intended_inputs) => {
-            entries_result = get_remote_links_and_load_entry_data(
+            entries_result = query_direct_remote_index_with_foreign_key(
                 intended_inputs, INTENT_BASE_ENTRY_TYPE,
                 INTENT_INPUT_OF_LINK_TYPE, INTENT_INPUT_OF_LINK_TAG,
             );
@@ -196,7 +196,7 @@ fn handle_query_processes(params: &QueryParams) -> ZomeApiResult<Vec<Response>> 
     };
     match &params.intended_outputs {
         Some(intended_outputs) => {
-            entries_result = get_remote_links_and_load_entry_data(
+            entries_result = query_direct_remote_index_with_foreign_key(
                 intended_outputs, INTENT_BASE_ENTRY_TYPE,
                 INTENT_OUTPUT_OF_LINK_TYPE, INTENT_OUTPUT_OF_LINK_TAG,
             );
