+++ happs/planning/zomes/fulfillment/code/src/fulfillment_requests.rs
@@ -18,11 +18,9 @@ use hdk_graph_helpers::{
         update_record,
         delete_record,
     },
-    links::{
-        get_links_and_load_entry_data,
-    },
     local_indexes::{
         create_direct_index,
+        query_direct_index_with_foreign_key,
     },
 };
 
@@ -148,14 +146,14 @@ fn handle_query_fulfillments(params: &QueryParams) -> ZomeApiResult<Vec<Response
     // :TODO: proper search logic, not mutually exclusive ID filters
     match &params.fulfills {
         Some(fulfills) => {
-            entries_result = get_links_and_load_entry_data(fulfills, COMMITMENT_FULFILLEDBY_LINK_TYPE, COMMITMENT_FULFILLEDBY_LINK_TAG);
+            entries_result = query_direct_index_with_foreign_key(fulfills, COMMITMENT_FULFILLEDBY_LINK_TYPE, COMMITMENT_FULFILLEDBY_LINK_TAG);
         },
         _ => (),
     };
     // :TODO: observation DNA handles this. Should queries be possible in planning DNA, too?
     // match &params.fulfilled_by {
     //     Some(fulfilled_by) => {
-    //         entries_result = get_links_and_load_entry_data(fulfilled_by, EVENT_FULFILLS_LINK_TYPE, EVENT_FULFILLS_LINK_TAG);
+    //         entries_result = query_direct_index_with_foreign_key(fulfilled_by, EVENT_FULFILLS_LINK_TYPE, EVENT_FULFILLS_LINK_TAG);
     //     },
     //     _ => (),
     // };
