+++ happs/planning/zomes/satisfaction/code/src/satisfaction_requests.rs
@@ -26,10 +26,8 @@ use hdk_graph_helpers::{
     rpc::{
         read_from_zome,
     },
-    links::{
-        get_links_and_load_entry_data,
-    },
     local_indexes::{
+        query_direct_index_with_foreign_key,
         create_direct_index,
     },
 };
@@ -184,13 +182,13 @@ fn handle_query_satisfactions(params: &QueryParams) -> ZomeApiResult<Vec<Respons
     // :TODO: implement proper AND search rather than exclusive operations
     match &params.satisfies {
         Some(satisfies) => {
-            entries_result = get_links_and_load_entry_data(satisfies, INTENT_SATISFIEDBY_LINK_TYPE, INTENT_SATISFIEDBY_LINK_TAG);
+            entries_result = query_direct_index_with_foreign_key(satisfies, INTENT_SATISFIEDBY_LINK_TYPE, INTENT_SATISFIEDBY_LINK_TAG);
         },
         _ => (),
     };
     match &params.satisfied_by {
         Some(satisfied_by) => {
-            entries_result = get_links_and_load_entry_data(satisfied_by, COMMITMENT_SATISFIES_LINK_TYPE, COMMITMENT_SATISFIES_LINK_TAG);
+            entries_result = query_direct_index_with_foreign_key(satisfied_by, COMMITMENT_SATISFIES_LINK_TYPE, COMMITMENT_SATISFIES_LINK_TAG);
         },
         _ => (),
     };
